refactor: 체스보드 출력 로직을  dto를 사용해 간소화
