+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,19 +17,6 @@ class ChessBoardTest {
     void create() {
         assertThatCode(ChessBoardFactory::makeChessBoard)
                 .doesNotThrowAnyException();
-    }
-
-    @Test
-    @DisplayName("현재 체스판 위에 존재하는 모든 기물을 구할 수 있다.")
-    void findAllPieces() {
-        Map<Position, Piece> positionPiece = new LinkedHashMap<>();
-        Position sourcePosition = Position.of(File.A, Rank.ONE);
-        positionPiece.put(sourcePosition, Queen.of(Color.BLACK));
-        ChessBoard chessBoard = new ChessBoard(positionPiece, new PawnMovementRule());
-
-        List<Piece> allPieces = chessBoard.findAllPieces();
-
-        assertThat(allPieces).hasSize(1);
     }
 
     @Test
